requirements.txt
# Core cryptography library
cryptography>=41.0.0

# QR code generation
qrcode[pil]>=7.4.0
Pillow>=10.0.0

# Optional: QR code reading (for scanning)
pyzbar>=0.1.9

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0